Remove dependency on `System.Web` in unit tests
The description of LightResize states that it is to be used in non-
ASP.NET context. For this reason, it appears somewhat inconsistent
to have a unit test for it that parses query strings.

This commit changes the unit test in question to no longer depend on
a query string. Then the dependency on `System.Web` can be removed.